yolo-garbage-service/requirements.txt
# Inference API + optional Streamlit demo
ultralytics==8.2.77
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
Pillow>=10.0.0
opencv-python-headless>=4.8.0
streamlit>=1.28.0
numpy>=1.24.0
